day19:find the error, refactored more
There was no reason to check if the error after finding the transformation.
The problem was a coincidence in the input that the pair happened twice and I was checking it with np.where(x > 1) but I should have check it with np.where(x >= 11) since I know it will occur at least 11 times.